requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
boto3==1.20.23
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.7
click==8.0.3
configparser==5.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.8
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
jwcrypto==1.0
kappa==0.6.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
pep517==0.12.0
pip-tools==6.4.0
placebo==0.10.0
psycopg2-binary==2.9.1
pycparser==2.21
pyparsing==3.0.3
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
tqdm==4.62.3
troposphere==3.1.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
wsgi-request-logger==0.4.6
zappa==0.54.1
